split demos
